requirements.txt
numpy #==1.18.5
cython #==0.29.21
librosa #==0.8.0
unidecode #==1.1.1
torch #==1.7.1
torchvision #==0.8.2
phonemizer #==2.2.1
pyopenjtalk
soundfile

google-cloud-storage
google-cloud-firestore